# unicorn-api/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateOrganizationRequest:
    properties:
      name:
        description: |-
          The name of the organization
          example: Acme Corporation
        type: string
    required:
    - name
    type: object
  handlers.CreateOrganizationResponse:
    properties:
      message:
        description: |-
          Success message
          example: Organization created successfully
        type: string
      organization:
        allOf:
        - $ref: '#/definitions/models.Organization'
        description: The created organization
    type: object
...
  handlers.CreateRoleResponse:
    properties:
      message:
        description: |-
          Success message
          example: Role created successfully
        type: string
      role:
        allOf:
        - $ref: '#/definitions/models.Role'
        description: The created role
    type: object
  handlers.CreateUserRequest:
    properties:
      email:
        description: |-
          The email address of the user (must be unique)
          example: john.doe@example.com
        type: string
      name:
        description: |-
          The display name of the user
          example: John Doe
        type: string
      password:
        description: |-
          The password for the user account (will be hashed)
          example: securePassword123
        minLength: 8
        type: string
      role_id:
        description: |-
          The ID of the role to assign to the user
          example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    required:
    - email
    - name
    - password
    - role_id
    type: object
  handlers.CreateUserResponse:
    properties:
      account:
        allOf:
        - $ref: '#/definitions/models.Account'
        description: The created user account
      message:
        description: |-
          Success message
          example: User created successfully
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      details:
        description: |-
          Additional error details
          example: Field 'email' is required
        type: string
      error:
        description: |-
          The error message
          example: Invalid request
        type: string
      status_code:
        description: |-
          The HTTP status code
          example: 400
        type: integer
      timestamp:
        description: |-
          The timestamp when the error occurred
          example: 2024-01-01T12:00:00Z
        type: string
    type: object
  handlers.GetOrganizationsResponse:
    properties:
      organization_name:
        type: string
      users:
        items:
          $ref: '#/definitions/handlers.OrganizationUser'
        type: array
    type: object
  handlers.GetRolesResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
    type: object
...
  handlers.LoginRequest:
    properties:
      email:
        description: |-
          The email address of the user
          example: john.doe@example.com
        type: string
      password:
        description: |-
          The password for the user account
          example: securePassword123
        type: string
    required:
    - email
    - password
    type: object
  handlers.LoginResponse:
    properties:
      expires_at:
        description: |-
          The expiration time of the token
          example: 2024-01-01T12:00:00Z
        type: string
      message:
        description: |-
          Success message
          example: Login successful
        type: string
      token:
        description: |-
          The JWT token for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJhY2NvdW50X2lkIjoiMTIzZTQ1NjctZTg5Yi0xMmQzLWE0NTYtNDI2NjE0MTc0MDAwIiwicm9sZV9pZCI6IjEyM2U0NTY3LWU4OWItMTJkMy1hNDU2LTQyNjYxNDE3NDAwMCIsImV4cCI6MTcwNDE2ODAwMH0.example_signature
        type: string
      token_type:
        description: |-
          The type of token
          example: Bearer
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@unicorn-api.com
    name: Unicorn API Support
    url: https://github.com/your-org/unicorn-api
  description: A comprehensive RESTful API for Unicorn services providing Identity
    and Access Management (IAM) functionality. This API supports user authentication,
    role-based access control, organization management, and JWT token handling.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Unicorn API
  version: "1.0"
paths:
...
  /api/v1/debug/token:
    get:
      consumes:
      - application/json
      description: Returns a static API token with admin permissions for testing purposes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get a debug API token for testing
      tags:
      - IAM
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Login
      tags:
      - IAM
  /api/v1/organizations:
    get:
      consumes:
      - application/json
      description: Returns the organization name and all users (name, role ID) in
        it
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetOrganizationsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the user's organization and its users
      tags:
      - IAM
    post:
      consumes:
      - application/json
      description: Create a new organization that can contain multiple accounts
      parameters:
      - description: Organization information
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateOrganizationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new organization
      tags:
      - IAM
  /api/v1/organizations/{org_id}/users:
    post:
      consumes:
      - application/json
      description: Create a user account in a specific organization with the specified
        role
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a user in an organization
      tags:
      - IAM
  /api/v1/roles:
    get:
      consumes:
      - application/json
      description: Returns all roles (name, permissions) for the authenticated user's
        organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetRolesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all roles in the user's organization
      tags:
      - IAM
    post:
      consumes:
      - application/json
      description: 'Create a new role with specified permissions. Permissions are:
        0=Read, 1=Write, 2=Delete'
      parameters:
      - description: Role information
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create a new role
      tags:
      - IAM
  /api/v1/roles/assign:
    post:
      consumes:
      - application/json
      description: Assign a role to a user or bot account
      parameters:
      - description: Role assignment information
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/handlers.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AssignRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Assign a role to an account
      tags:
      - IAM
...
  /api/v1/token/refresh:
    post:
      consumes:
      - application/json
      description: Refresh an expired JWT token
      parameters:
      - description: Refresh token request
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Refresh JWT token
      tags:
      - IAM
  /api/v1/token/validate:
    get:
      consumes:
      - application/json
      description: Validate a JWT token and return its claims
      parameters:
      - description: JWT token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ValidateTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Validate JWT token
      tags:
      - IAM
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"